isolated-docker/requirements.txt
Django==2.0.3
djangorestframework==3.7.7
dj-database-url==0.4.2
psycopg2==2.7.3.2